Bot/requirements.txt
python-telegram-bot==20.4
ccxt==3.1.56
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
python-dotenv==1.0.0